+++ R/aggregates.R
@@ -146,18 +146,20 @@ primary_aggregates <- function(.sutdata = NULL,
 #' Final demand energy and exergy aggregates
 #'
 #' Calculates aggregate final demand energy from a data frame of Supply-Use matrices.
-#' The calculation includes non-energy uses if they are present in the final demand matrix.
-#' The calculation does not include balancing items (Losses and Statistical differences).
-#' If `.sutdata` is a data frame, `fd_sectors` should be the name of a column in the data frame.
+#' The calculation counts only `fd_sectors` in final demand aggregates.
 #' If `.sutdata` is `NULL`, `fd_sectors` can be a single vector of industry names.
+#'
+#' Net energy demand is calculated by `matsbyname::sum_byname(Y_fd)`, with sums across rows, columns, or total as needed.
+#' Gross energy demand is calculated by `matsbyname::sum_byname(Y_fd) + matsbyname::sum_byname(U_EIOU)`,
+#' with sums across rows, columns, or total as needed.
 #'
 #' @param .sutdata A data frame with columns of matrices from a supply-use analysis.
 #' @param fd_sectors A vector of names of sectors in final demand.
 #'                   Names should include columns in the `Y` and `U_EIOU` matrices
 #'                   to cover both net (in `Y`) and gross (in `Y` and `U_EIOU`) final demand.
 #' @param piece,notation,pattern_type,prepositions Arguments that control the way row and column matching
-#'                                                 is accomplished when selecting parts of the **R**, **V**, and **Y**
-#'                                                 matrices for primary aggregation.
+#'                                                 is accomplished when selecting parts of the **U_EIOU** and **Y**
+#'                                                 matrices for final demand aggregation.
 #'                                                 These arguments are passed to
 #'                                                 `matsbyname::select_rowcol_piece_byname()` and eventually
 #'                                                 `RCLabels::match_by_pattern()` and
@@ -167,14 +169,12 @@ primary_aggregates <- function(.sutdata = NULL,
 #'                                                 `notation = RCLabels::notations_list`,
 #'                                                 `pattern_type = "exact"`, and
 #'                                                 `prepositions = RCLabels::prepositions_list`.
-#' @param U,U_feed,Y Input matrices. See `Recca::psut_cols`.
+#' @param U_eiou,Y Input matrices. See `Recca::psut_cols`.
 #' @param by One of "Product", "Sector", or "Total" to indicate the desired aggregation:
 #'           "Product" for aggregation by energy carrier (Crude oil, Primary solid biofuels, etc.),
 #'           "Sector" for aggregation by final demand sector (Agriculture/forestry, Domestic navigation, etc.), or
 #'           "Total" for aggregation over both Product and Sector (the default).
 #' @param net_aggregate_demand,gross_aggregate_demand See `Recca::aggregate_cols`.
-#'        Net energy demand is calculated by `matsbyname::sumall_byname(Y_fd)`.
-#'        Gross energy demand is calculated by `matsbyname::sumall_byname(Y_fd) + matsbyname::sumall_byname(U_EIOU)`.
 #'        Defaults are `Recca::aggregate_cols$net_aggregate_demand` and
 #'        `Recca::aggregate_cols$gross_aggregate_demand`.
 #'
@@ -200,10 +200,9 @@ finaldemand_aggregates <- function(.sutdata = NULL,
                                    pattern_type = c("exact", "leading", "trailing", "anywhere", "literal"),
                                    prepositions = RCLabels::prepositions_list,
                                    # Input names
-                                   U = Recca::psut_cols$U,
-                                   U_feed = Recca::psut_cols$U_feed,
+                                   U_eiou = Recca::psut_cols$U_eiou,
                                    Y = Recca::psut_cols$Y,
-                                   by = c("Total", "Product", "Sector"),
+                                   by = c("Total", "Product", "Sector", "Industry"),
                                    # Output names
                                    net_aggregate_demand = Recca::aggregate_cols$net_aggregate_demand,
                                    gross_aggregate_demand = Recca::aggregate_cols$gross_aggregate_demand) {
@@ -211,8 +210,7 @@ finaldemand_aggregates <- function(.sutdata = NULL,
   pattern_type <- match.arg(pattern_type)
   by <- match.arg(by)
 
-  fd_func <- function(U_mat, U_feed_mat, Y_mat){
-    U_eiou_mat <- matsbyname::difference_byname(U_mat, U_feed_mat)
+  fd_func <- function(U_eiou_mat, Y_mat){
 
     Y_mat_cols <- Y_mat %>%
       matsbyname::select_rowcol_piece_byname(retain = fd_sectors,
@@ -242,7 +240,7 @@ finaldemand_aggregates <- function(.sutdata = NULL,
       gross <- U_eiou_mat_cols %>%
         matsbyname::rowsums_byname() %>%
         matsbyname::sum_byname(net)
-    } else if (by == "Sector") {
+    } else if (by %in% c("Sector", "Industry")) {
       net <- Y_mat_cols %>%
         matsbyname::colsums_byname()
       gross <- U_eiou_mat_cols %>%
@@ -270,7 +268,7 @@ finaldemand_aggregates <- function(.sutdata = NULL,
     }
     list(net, gross) %>% magrittr::set_names(c(net_aggregate_demand, gross_aggregate_demand))
   }
-  matsindf::matsindf_apply(.sutdata, FUN = fd_func, U_mat = U, U_feed_mat = U_feed, Y_mat = Y)
+  matsindf::matsindf_apply(.sutdata, FUN = fd_func, U_eiou_mat = U_eiou, Y_mat = Y)
 }